+++ scheduler/conf.c
@@ -1975,8 +1975,7 @@ get_addr_and_mask(const char *value,	/* I - String from config file */
                  mask + 3) != 4)
         return (0);
 
-      mask[3] |= ((unsigned)mask[0] << 24) | ((unsigned)mask[1] << 16) |
-                 ((unsigned)mask[2] << 8);
+      mask[3] |= (mask[0] << 24) | (mask[1] << 16) | (mask[2] << 8);
       mask[0] = mask[1] = mask[2] = 0;
     }
     else
@@ -3896,10 +3895,10 @@ read_policy(cups_file_t *fp,		/* I - Configuration file */
             char        *policy,	/* I - Location name/path */
 	    int         linenum)	/* I - Current line number */
 {
-  int			i;		/* Looping var */
+  size_t			i;		/* Looping var */
   cupsd_policy_t	*pol;		/* Policy */
   cupsd_location_t	*op;		/* Policy operation */
-  int			num_ops;	/* Number of IPP operations */
+  size_t			num_ops;	/* Number of IPP operations */
   ipp_op_t		ops[100];	/* Operations */
   char			line[HTTP_MAX_BUFFER],
 					/* Line buffer */
@@ -3965,7 +3964,7 @@ read_policy(cups_file_t *fp,		/* I - Configuration file */
 	if (*valptr)
 	  *valptr++ = '\0';
 
-        if (num_ops < (int)(sizeof(ops) / sizeof(ops[0])))
+        if (num_ops < (sizeof(ops) / sizeof(ops[0])))
 	{
 	  if (!_cups_strcasecmp(value, "All"))
 	    ops[num_ops ++] = IPP_ANY_OPERATION;
